misc. sensors

+++ Automation/OpticalCamExamples/SimpleLive_Pyueye_OpenCV.py
@@ -166,7 +166,7 @@ a=20
 
 
 
-a=ueye.DOUBLE(40)
+a=ueye.DOUBLE(10)
 c=ueye.DOUBLE()
 b=ueye.UINT(8)
 thing=ueye.is_Exposure(hCam,ueye.IS_EXPOSURE_CMD_SET_EXPOSURE, a,b )
